Update Let and Const.js
in order to calculate a float result, we need to get input as float (decimal)

+++ Day 1/Let and Const.js
@@ -1,7 +1,7 @@
 function main() {
     // Write your code here. Read input using 'readLine()' and print output using 'console.log()'.
     const PI = Math.PI;
-    var r = parseInt(readLine());
+    var r = parseFloat(readLine());
     // Print the area of the circle:
     var area = PI * r * r;
     console.log(area);
